Test for bool and 'long long'.  This allows xosview to be built by HP CC (which is needed to run purify).

+++ timeval.h
@@ -17,14 +17,14 @@
 class TimeVal {
 public:
   TimeVal(unsigned long sec = 0, unsigned long usec = 0) {
-    _val.tv_sec = sec;
+    _val.tv_sec = (int)sec;
     _val.tv_usec = usec;
   }
   TimeVal(const struct timeval &val) { _val = val; }
 
   unsigned long sec(void) const { return _val.tv_sec; }
   unsigned long usec(void) const { return _val.tv_usec; }
-  void sec(unsigned long s) { _val.tv_sec = s; }
+  void sec(unsigned long s) { _val.tv_sec = (int)s; }
   void usec(unsigned long us) { _val.tv_usec = us; }
 
   operator struct timeval(void) const { return _val; }
